requirements.txt
# Core Dependencies
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0

# Database
pymongo>=4.0.0
motor>=3.0.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
